Update README.md
In the final project of the laboratory for analog VLSI it was requested to build an ADC that is read serially. Given that the system input is an analog signal and the system output (BUS) is a signal. It was also requested to synchronize the signals and switches so that it would be possible to read the entire ENCODER output before moving to the next voltage level. The input frequency to the system is a sinusoidal signal around f=100K(Hz).